Extract code into own file
